Add a callback named SummaryCollector and delete SummaryStep callback
I added a SummaryCollector to help users automatically collect information
such as the network, loss, learning rate and so on, making it easier to collect this information.

It also can collect train lineage and eval lineage information which is
collected by TrainLineage Callback and EvalLineage Callback in
MindInsight.

I also add some UT for SummaryCollect to keep the code correct.