add textrcnn into model zoo
Signed-off-by: zymaa <317958662@qq.com>

TextRCNN, a model for text classification, which is proposed by the Chinese Academy of Sciences in 2015.
TextRCNN actually combines RNN and CNN, first uses bidirectional RNN to obtain upper semantic and grammatical information of the input text,
and then uses maximum pooling to automatically filter out the most important feature.

Signed-off-by: zymaa <317958662@qq.com>