support env_config_path to set the config file for DFX
1. add env_config_path for ms_context
2. add EnvConfigParser to parse config
3. add in CMakeLists
4. clang-format, cpplint
5. fix spelling in ms_context

Use EnvConfigParser in the constructed function of base_recorder,
when a recorder is initialized, its direcctory will be set as the
value from config file. In addition, recorder_manager will get the
'enable' from config file. If 'enable' is true and the path is not a
valid value, it will be the default value: './rdr'.