Refactor GraphKernelExpander (2nd submission)
Changed the expander to class, and used the class name to identify the operator.
Moved the original op logic into the `_expand` function; added a `_check` function to check op inputs.

Use decorator to register the whitelist formats that operator supports,
and the decorator will change the `_check` function to check the formats.