add MS_RDR_ENABLE and MS_RDR_PATH
MS_RDR_ENABLE should be in {0, 1}. if it is set as other value, it will
be 0 and WARNING information will be printed to tell user.
The settings in config file set by the user takes precedence over the
environment variables. If 'rdr_enable' is 0 in environment variable,
and then 'enable' is true in config file, the RDR will turned on.