Bugfix for DropoutExpander
When creating the new GKDropout, the original process used the Dropout's input's kernel_info to infer the kernel_info of new node.
but if the input is a Tensor, then it hasn't kernel_info, which causes the GKDropout's input_kernel_info to be invalid.
In this submit, I manually set the kernel_infos to ensure they are always valid.