Bugfix in graph_kernel_cluster when the op connects to UpdateState
In PR !16073, the tail inputs of UpdateState is ignored in TransformSegmentToAnfGraph,
but we spread the inputs of UpdateState in the pass SpreadUpdateState, so the error occurs.

In this submission, we call ShrinkUpdateState before clustering operators and
call SpreadUpdateState again after that, to avoid the problem.
The ParallelFusion also calls the TransformSegmentToAnfGraph interface, so we change it together.

Note: this submission is a temporary solution, we will rewrite the TransformSegmentToAnfGraph in
graphkernel module, without these special processes.

To speed up the pass SpreadUpdateState, we create new UpdateState node and use "mng.Replace",
instead of setting inputs and use "mng.KeepRoots".