Save and restore the output abstract of original graph
For convenience, we may change some operators' shape in equivalent way,
such as changing the scalar value (shape is empty) to a tensor with shape [1].

It's ok for the intermediate tensors, but not for the outputs.

So we save the output shapes in pre-process stage, and restore them in post-process stage.