GraphKernel supports CPU
only Linux system is supported now.

change the default value of `ENABLE_AKG` to off, and controlled by option `-K`.
the `ENABLE_AKG` is auto enabled when `ENABLE_GPU` or `ENABLE_D` is on.
since now, we can use `ENABLE_AKG` to control the compilation of graphkernel
and akg codes.

fix usage description for option "-K", it should be "[-K on|off]".

LLVM is required by akg for cpu kernels, so AKG for cpu is default disabled now.