Move InsertCastCpu before BuildKernel
It's necessary to immediately insert cast for the nodes that dtype was changed in SelectKernel,
instead of after BuildKernel. otherwise, some fusion passes cannot work as expected,
like Graph-Kernel optimization.

By now, the shape of LSTM's reserved output will be changed in InitKernel (called by BuildKernel),
and this output is only used by its gradient operator, so we can ignore the reserved output in this pass.