Add "is_internal_output" attribute for related nodes
The internal output nodes of KernelGraph are stored in a `map`
for memory reuse between KernelGraphs.

However, the nodes may be changed in graphkernel optimization.
the passes of graphkernel are base on `FuncGraph`, instead of `KernelGraph`,
so we cannot use the `ReplaceInternalOutput` interface.

This commit, we add an attribute for these nodes and skip them in graphkernel's passes.