Check nullptr in the 'Symbol::as' interface
To make it convenience to use the symbol object, change the original 'as' interface to checking result.
If the object is not the subclass of given class, an exception will be thrown, so the user can use the
symbol without step-by-step checking, like 'shape->as<ListSymbol>()->item_as<IntSymbol>(0)->value()'.

The original behavior of 'as' interface is changed to 'as_noexcept', which returns nullptr when class unmatched.